+++ CHANGELOG.md
@@ -11,7 +11,11 @@
 - Help buffers: look up the currently used keybinding for each available function, instead of using
   hardcoded bindings which may have been modified by the user.
 
-- CockroachDB: show information on the configured users in the “More backend information” buffer.
+- CockroachDB variant: show information on the configured users in the “More backend information”
+  buffer.
+
+- Risingwave variant: improve the “More backend information” buffer, and implement the “display
+  functions” feature.
 
 
 ## [0.26] - 2025-07-09